+++ vfl/util/grid.h
@@ -10,23 +10,19 @@
 
 int grid_validate (const Matrix *grid);
 
-unsigned int grid_dims (const Matrix *grid);
+size_t grid_dims (const Matrix *grid);
 
-unsigned int grid_elements (const Matrix *grid);
+size_t grid_elements (const Matrix *grid);
 
 int grid_iterator_alloc (const Matrix *grid,
-                         unsigned int *elems,
-                         unsigned int **idx,
-                         unsigned int **sz,
+                         size_t *elems,
+                         size_t **idx,
+                         size_t **sz,
                          Vector **x);
 
-void grid_iterator_free (unsigned int *idx,
-                         unsigned int *sz,
-                         Vector *x);
+void grid_iterator_free (size_t *idx, size_t *sz, Vector *x);
 
-int grid_iterator_next (const Matrix *grid,
-                        unsigned int *idx,
-                        unsigned int *sz,
+int grid_iterator_next (const Matrix *grid, size_t *idx, size_t *sz,
                         Vector *x);
 
 #endif /* !__VFL_GRID_H__ */
